appcinema/api/serializers.py
from rest_framework import serializers

from .models import Movie, Booking


class MovieSerializer(serializers.ModelSerializer):

    class Meta:
        model = Movie
        fields = ('id', 'name', 'date', '__str__')


class BookingSerializer(serializers.ModelSerializer):

    class Meta:
        model = Booking
        fields = ('id', 'movie', 'user',)
