requirements.txt
amqp==2.1.4
appdirs==1.4.3
asn1crypto==0.22.0
billiard==3.5.0.2
celery==4.0.2
cffi==1.10.0
cryptography==1.8.1
defusedxml==0.4.1
Django==1.10.6
django-debug-toolbar==1.7
django-extensions==1.7.7
djangorestframework==3.6.2
idna==2.5
kombu==4.0.2
ndg-httpsclient==0.4.2
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pusher==1.6
pyasn1==0.2.3
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
sqlparse==0.2.3
urllib3==1.17
vine==1.1.3
Werkzeug==0.12.1
